workspace/requirements.txt
# Combined Requirements for Adobe AI Image Editing Pipeline
# This file contains all dependencies needed across all tasks

# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
onnxruntime[GPU]
# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
rembg>=2.0.50

# Model-specific (optional but recommended for performance)
# Uncomment the line below if you have CUDA GPU support
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118

# Background removal
# rembg dependencies are included above
